Add Project 1&2
This is the upload of the working code for Project 2. It is a course program that allows users to load, search, and display courses and their prerequisites. 
This upload also includes Project 1, which is a reflection on Hash, Vectors, and Trees. It includes runtimes and pseudocode for each type to display and search courses.